shortcut-fe: fix min() distinct pointer type error on new GCC
Cast mac_dev->addr_len (unsigned char) to int so both operands of the
min() macro have the same type, avoiding 'comparison of distinct
pointer types' being promoted to an error by newer GCC with -Werror.

+++ package/extra/shortcut-fe/shortcut-fe/src/sfe_cm.c
@@ -289,7 +289,7 @@ skip_dst_lookup:
 		goto ret_fail;
 	}
 
-	memcpy(mac_addr, neigh->ha, (size_t)min(mac_dev->addr_len, ETH_ALEN));
+	memcpy(mac_addr, neigh->ha, (size_t)min((int)mac_dev->addr_len, ETH_ALEN));
 
 	dev_hold(mac_dev);
 	*dev = mac_dev;
